:pushpin: Alterando a versão do dotnet

+++ SalesWebMvc/Controllers/DepartamentsController.cs
@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.Rendering;
+﻿using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using SalesWebMvc.Data;
 using SalesWebMvc.Models;
@@ -22,13 +17,15 @@ namespace SalesWebMvc.Controllers
         // GET: Departaments
         public async Task<IActionResult> Index()
         {
-            return View(await _context.Departament.ToListAsync());
+              return _context.Departament != null ? 
+                          View(await _context.Departament.ToListAsync()) :
+                          Problem("Entity set 'SalesWebMvcContext.Departament'  is null.");
         }
 
         // GET: Departaments/Details/5
         public async Task<IActionResult> Details(int? id)
         {
-            if (id == null)
+            if (id == null || _context.Departament == null)
             {
                 return NotFound();
             }
@@ -50,8 +47,8 @@ namespace SalesWebMvc.Controllers
         }
 
         // POST: Departaments/Create
-        // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
-        // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
+        // To protect from overposting attacks, enable the specific properties you want to bind to.
+        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("Id,Name")] Departament departament)
@@ -68,7 +65,7 @@ namespace SalesWebMvc.Controllers
         // GET: Departaments/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
-            if (id == null)
+            if (id == null || _context.Departament == null)
             {
                 return NotFound();
             }
@@ -82,8 +79,8 @@ namespace SalesWebMvc.Controllers
         }
 
         // POST: Departaments/Edit/5
-        // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
-        // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
+        // To protect from overposting attacks, enable the specific properties you want to bind to.
+        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("Id,Name")] Departament departament)
@@ -119,7 +116,7 @@ namespace SalesWebMvc.Controllers
         // GET: Departaments/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
-            if (id == null)
+            if (id == null || _context.Departament == null)
             {
                 return NotFound();
             }
@@ -139,15 +136,23 @@ namespace SalesWebMvc.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            if (_context.Departament == null)
+            {
+                return Problem("Entity set 'SalesWebMvcContext.Departament'  is null.");
+            }
             var departament = await _context.Departament.FindAsync(id);
-            _context.Departament.Remove(departament);
+            if (departament != null)
+            {
+                _context.Departament.Remove(departament);
+            }
+            
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
         }
 
         private bool DepartamentExists(int id)
         {
-            return _context.Departament.Any(e => e.Id == id);
+          return (_context.Departament?.Any(e => e.Id == id)).GetValueOrDefault();
         }
     }
 }
